\name{AppResults-class}
\Rdversion{1.1}

\docType{class}

\alias{appResultItem}
\alias{appResultItem-class}
\alias{appResultCollection}
\alias{appResultCollection-class}

\alias{AppResultsSummary-class}
\alias{AppResults-class}
\alias{AppResults}
\alias{AppResults,missing-method}
\alias{AppResults,AppAuth-method}
\alias{AppResults,AppResultsSummary-method}

\alias{listAppResults}
\alias{listAppResults,AppAuth-method}
\alias{listAppResults,Projects-method}

\alias{createAppResults}
\alias{createAppResults,AppAuth-method}
\alias{createAppResults,Projects-method}


\title{\code{AppResults} and \code{"AppResultsSummary"} objects}

\description{
  Classes and methods to handle the AppResults resource.
}

\references{
  https://developer.basespace.illumina.com/docs/content/documentation/rest-api/api-reference
}

\author{
  Adrian Alexa
}


\seealso{
  \code{\linkS4class{AppAuth}} and \code{\linkS4class{Projects}}.
}
\examples{
showClass("AppAuth")
}
\keyword{classes}
\keyword{methods}
