\name{Miscs}
\alias{objectSize}
\alias{messageHandler}
\title{

  Miscs functions ... 

}
\description{

  \code{objectSize} .....

  \code{messageHandler}
}

\usage{
objectSize(x, units = "auto")
messageHandler(logfile)
}
\arguments{

  \item{x}{\code{xxx} ...}
  \item{units}{\code{character()} same as for \code{object.size}}
  \item{logfile}{\code{character()} name of the log file. Can be omitted}
}

\details{

  The \code{objectSize()} function is generally
}

\value{
  \code{messageHandler()} returns a function ...
}
\seealso{

  \code{\link{object.size}} ...

}

