\name{Private methods}
\Rdversion{1.1}

\docType{class}

\alias{show,staticHref-method}
\alias{show,Href-method}

\alias{auth,Href-method}

\alias{as.character,staticHref-method}

\title{Internal methods}

\description{
  This page is used as a placeholder for private methods in order to
  pass the R checks.
}

\details{
  Internal use.
}

\author{
  Adrian Alexa
}

\keyword{classes}
\keyword{methods}
