\name{AppSessions-class}
\Rdversion{1.1}

\docType{class}

\alias{AppSessions-class}
\alias{AppSessions}
\alias{AppSessions,AppAuth-method}
\alias{AppSessions,AppResults-method}

\alias{appSessionItem}
\alias{appSessionItem-class}

\alias{listAppSessions}
\alias{listAppSessions,AppAuth-method}

\alias{updateAppSessions}
\alias{updateAppSessions,AppAuth-method}


\title{\code{AppSessions} object}

\description{
  Classes and methods to handle the AppSessions resource.
}

\details{
  coming soon... 
}

\references{
  https://developer.basespace.illumina.com/docs/content/documentation/rest-api/api-reference
}

\author{
  Adrian Alexa
}


\seealso{
  \code{\linkS4class{AppAuth}} and \code{\linkS4class{Response}}.
}
\examples{
showClass("AppAuth")
}
\keyword{classes}
\keyword{methods}
